Day 7 added
Till I can finish my work on the code, the original json file changed in the website. The measurement method was different. But I tried to optimize the data so that I can get 14 day incidence values per week. Clock charts were established before the changes in the original data.